requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
pydantic==2.13.4
pydantic-extra-types==2.11.1
pydantic-settings==2.11.0
python-dotenv==1.2.1
python-multipart==0.0.20
veryfi==5.0.1
requests==2.32.5
httpx==0.28.1
email-validator==2.3.0
nltk==3.9.2
numpy==2.0.2
pytest==8.4.2
pytest-cov==6.x.x
